Improved exception handling

+++ Business/Groups/GroupManager.cs
@@ -2,12 +2,14 @@
 using Business.Models;
 using System.Linq;
 using System;
+using Business.Errors;
 
 namespace Business.Groups
 {
     class GroupManager : IGroupManager
     {
         private readonly IGroupRepository groupRepository;
+
         public GroupManager(IGroupRepository groupRepository)
         {
             this.groupRepository = groupRepository;
@@ -17,9 +19,14 @@ namespace Business.Groups
         {
             var loggedInUser = groupRepository.GetGroupUser(groupId, callerId);
 
-            //Iterates through all user ids, creates a GroupUserModel and sends that to our group repository
+            if (loggedInUser == null)
+            {
+                throw new InvalidArgumentException("The user is not in the specified group", ParamNameType.GroupId);
+            }
+
             if (loggedInUser.IsAdministrator)
             {
+                //Iterates through all user ids, creates a GroupUserModel and sends that to our group repository
                 groupRepository.AddMembersToGroup(groupUserModel: userIds.Select(userId => new GroupUserModel()
                 {
                     Group = new GroupModel()
@@ -34,7 +41,7 @@ namespace Business.Groups
             }
             else
             {
-                throw new ArgumentException("The user was not authorized to add users to the group", "callerId");
+                throw new InvalidArgumentException("The user was not authorized to add users to the group", ParamNameType.CallerId);
             }
         }
 
@@ -73,13 +80,18 @@ namespace Business.Groups
         {
             var loggedInUser = groupRepository.GetGroupUser(groupId, callerId);
 
+            if (loggedInUser == null)
+            {
+                throw new InvalidArgumentException("The user is not in the specified group", ParamNameType.GroupId);
+            }
+
             if (loggedInUser.IsAdministrator)
             {
                 groupRepository.DeleteGroup(groupId);
             }
             else
             {
-                throw new ArgumentException("The user was not authorized to delete the group", "callerId");
+                throw new InvalidArgumentException("The user was not authorized to delete the group", ParamNameType.CallerId);
             }
         }
 
@@ -87,16 +99,30 @@ namespace Business.Groups
         {
             var loggedInUser = groupRepository.GetGroupUser(groupId, callerId);
 
+            if (loggedInUser == null)
+            {
+                throw new InvalidArgumentException("The user is not in the specified group", ParamNameType.GroupId);
+            }
+
             if (loggedInUser.IsAdministrator)
             {
                 groupRepository.UpdateGroup(groupId, groupName, callerId);
+            }
+            else
+            {
+                throw new InvalidArgumentException("The user was not authorized to update the group", ParamNameType.CallerId);
             }
         }
 
         public void RemoveUsersFromGroup(int groupId, List<int> userIds, int callerId)
         {
             var loggedInUser = groupRepository.GetGroupUser(groupId, callerId);
-            //Iterates through all user ids, creates a GroupUserModel and sends that to our group repository
+
+            if (loggedInUser == null)
+            {
+                throw new InvalidArgumentException("The user is not in the specified group", ParamNameType.GroupId);
+            }
+
             if (loggedInUser.IsAdministrator)
             {
                 //Iterates through all user ids, creates a GroupUserModel and sends that to our group repository
@@ -114,13 +140,18 @@ namespace Business.Groups
             }
             else
             {
-                throw new ArgumentException("The user was not authorized to remove users to the group", "callerId");
+                throw new InvalidArgumentException("The user was not authorized to remove users to the group", ParamNameType.CallerId);
             }
         }
 
         public void AddRolesToGroup(int groupId, int callerId, IEnumerable<int> roleIds)
         {
             var loggedInUser = groupRepository.GetGroupUser(groupId, callerId);
+
+            if (loggedInUser == null)
+            {
+                throw new InvalidArgumentException("The user is not in the specified group", ParamNameType.GroupId);
+            }
 
             if (loggedInUser.IsAdministrator)
             {
@@ -138,13 +169,18 @@ namespace Business.Groups
             }
             else
             {
-                throw new ArgumentException("The user was not authorized to add roles to the group", "callerId");
+                throw new InvalidArgumentException("The user was not authorized to add roles to the group", ParamNameType.CallerId);
             }
         }
 
         public void RemoveRolesFromGroup(int groupId, int callerId, IEnumerable<int> roleIds)
         {
             var loggedInUser = groupRepository.GetGroupUser(groupId, callerId);
+
+            if (loggedInUser == null)
+            {
+                throw new InvalidArgumentException("The user is not in the specified group", ParamNameType.GroupId);
+            }
 
             if (loggedInUser.IsAdministrator)
             {
@@ -162,13 +198,18 @@ namespace Business.Groups
             }
             else
             {
-                throw new ArgumentException("The user was not authorized to remove roles from the group", "callerId");
+                throw new InvalidArgumentException("The user was not authorized to remove roles from the group", ParamNameType.CallerId);
             }
         }
 
         public void SetUserAdministratorOnGroup(int groupId, int userId, int callerId, bool isAdministrator)
         {
             var loggedInUser = groupRepository.GetGroupUser(groupId, callerId);
+
+            if (loggedInUser == null)
+            {
+                throw new InvalidArgumentException("The user is not in the specified group", ParamNameType.GroupId);
+            }
 
             if (loggedInUser.IsAdministrator)
             {
@@ -188,7 +229,7 @@ namespace Business.Groups
             }
             else
             {
-                throw new ArgumentException("The user was not authorized to change administrator", "callerId");
+                throw new InvalidArgumentException("The user was not authorized to change administrator", ParamNameType.CallerId);
             }
         }
 
@@ -196,28 +237,24 @@ namespace Business.Groups
         
             var loggedInUser = groupRepository.GetGroupUser(groupId, callerId);
 
-            if (loggedInUser != null)
+            if (loggedInUser == null)
             {
-                return groupRepository.GetAllGroupUsers(groupId);
+                throw new InvalidArgumentException("The user is not in the specified group", ParamNameType.GroupId);
             }
-            else
-            {
-                throw new ArgumentException("Failed to get all group users", "callerId");
-            }
+
+            return groupRepository.GetAllGroupUsers(groupId);
         }
 
         public IEnumerable<UserModel> GetAllGroupAdmins(int groupId, int callerId) {
 
             var loggedInUser = groupRepository.GetGroupUser(groupId, callerId);
 
-            if (loggedInUser != null)
+            if (loggedInUser == null)
             {
-                return groupRepository.GetAllGroupAdmins(groupId);
+                throw new InvalidArgumentException("The user is not in the specified group", ParamNameType.GroupId);
             }
-            else
-            {
-                throw new ArgumentException("Failed to get all group admins", "callerId");
-            }
+
+            return groupRepository.GetAllGroupAdmins(groupId);
         }
 
         public IEnumerable<GroupModel> GetGroupsForUser(int userId)
